% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_sdc.R
\name{apply_sdc}
\alias{apply_sdc}
\title{Apply statistical disclosure control across a dataset}
\usage{
apply_sdc(data, level = 5, rounding = TRUE, round_val = 5, mask = -1)
}
\arguments{
\item{data}{a dataframe object}

\item{level}{level determines the value that SDC should be applied from.
for example, \code{level = 5} applies SDC to all values less than 5 and greater
than 0 (1, 2, 3, 4).}

\item{rounding}{\code{TRUE} or \code{FALSE}. determines if rounding of none SDC values
should be applied.}

\item{round_val}{determines what nearest value should be used when applying
rounding. for example \code{round_val = 5} will round values to the nearest 5}

\item{mask}{determines what the actual mask value is for SDC values. defaults
to \code{-1}. can be numeric, character, or \code{NA_real_}.}
}
\description{
This function allows the application of SDC for numeric columns within a
dataframe. the function evaluates based upon the raw value of the column
and doesn't handle any SDC required by inference.
}
\details{
natural 0s are always preserved.

This function can be ran as part of a dplyr pipe.
}
\examples{
raw_data$mumh_annual_stp_data \%>\% apply_sdc(rounding = F) \%>\% View
}
