% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_data.R
\name{available_data}
\alias{available_data}
\title{Get latest available periods of data}
\usage{
available_data(database)
}
\arguments{
\item{database}{String. Name of the DB to be passed to \code{nhsbsaR::con_nhsbsa}.
one of c("DALP", "DWCP")}
}
\value{
A \code{NULL} object. this function only sends a message to the console.
}
\description{
\code{available_data} returns a message to the console with the latest available
month, full financial quarter, and full financial year of data from the
NHSBSA Data Warehouse.
}
\examples{
available_data("DWCP")

#The latest month of data is:
#202111

#The latest complete financial quarter of data is:
#Jul - Sep 21

#The latest complete financial year of data is:
#2020/2021

}
