% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_chart_hc.R
\name{covid_chart_hc}
\alias{covid_chart_hc}
\title{Easy helper for covid chart in highcharter}
\usage{
covid_chart_hc(data, title = NULL)
}
\arguments{
\item{data}{a dataframe to be charted}

\item{title}{the chart title (defaults to NULL)}
}
\description{
Create covid model chart in highcharter
}
\examples{
covid_chart_hc()
}
