% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_model.R
\name{covid_model}
\alias{covid_model}
\title{Covid model function}
\usage{
covid_model(data)
}
\arguments{
\item{data}{The data to be input into the model}
}
\description{
This function creates a linear model using the lm() function to predict number
of items and patients for BNF chapters specified in input data. The Covid-19
period is defined as February 2020 onwards.
Prediction intervals are set at 90\%, 95\% and 99\%
}
\examples{
covid_model(data)
}
