% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\alias{create_tdim}
\alias{create_org_dim}
\alias{create_drug_dim}
\alias{create_fact}
\title{Create queries to extract data from NHSBSA DWH}
\usage{
create_tdim(
  con,
  start = 201504,
  end = sql("MGMT.PKG_PUBLIC_DWH_FUNCTIONS.f_get_latest_period('EPACT2')")
)

create_org_dim(con, country = 1)

create_drug_dim(con, bnf_codes, ...)

create_fact(con)
}
\arguments{
\item{con}{a database object created \code{nhsbsaR::con_nhsbsa} or \code{DBI}
equivalent}

\item{start}{year month of start of period in YYYYMM format. Defaults to
start of patient data in DWH, 201504}

\item{end}{year month of end of period in YYYYMM format. Defaults to latest
year month in DWH using in built \code{f_get_latest_period('EPACT2')} function}
}
\value{
a lazy table that can be used in \code{dplyr} chains.
}
\description{
A series of \verb{create_*()} functions that generate lazy db tables powered by
\code{dbplyr} that can chained into other \code{dplyr} functions to query the NHSBSA
Data warehouse in a flexible manner.
\itemize{
\item \code{create_tdim()} builds lazy table that queries \code{DIM.YEAR_MONTH_DIM} to give
latest complete financial quarter period.
\item \code{create_org_dim()} builds lazy table that queries \code{DIM.CUR_LEVEL_5_FLAT_DIM}.
Defaults to England only.
\item \code{create_drug_dim()} builds lazy table that queries \code{DIM.CDR_EP_DRUG_BNF_DIM}.
By default returns all BNF hierarchy information. Can be filtered using any
level of the BNF hierarchy. Additional fields can be added.
\item \code{create_fact()} builds lazy table that queries
\code{AML.PX_FORM_ITEM_ELEM_COMB_FACT_AV}. Returns Items and NIC by default with
patient ID level information.
}
}
\examples{
con <- nhsbsaR::con_nhsbsa(database = "DWCP")

df <- create_tdim(con, start = 201904) \%>\%
  create_org_dim(con, country = 1) \%>\%
  create_drug_dim(con, bnf_codes = c("0401", "0402", "0403")) \%>\%
  create_fact(con) \%>\%
  collect
}
