% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWB.R
\name{createWB}
\alias{createWB}
\title{Easy helper for 'createWB'}
\usage{
createWB(filepath, sheets)
}
\arguments{
\item{filepath}{the filepath to save the workbook to}

\item{sheets}{a list of sheet names require}
}
\description{
Create base workbook 'wb' and add sheets required
}
\examples{
sheetNames <- c("test1","test2","test3")
createWB("C:/test.xlsx", sheetNames)
}
