% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_column_to_numeric.R
\name{excel_column_to_numeric}
\alias{excel_column_to_numeric}
\title{Easy helper for 'Excel column to numeric'}
\usage{
excel_column_to_numeric(column_letter)
}
\arguments{
\item{column_letter}{the letter of the Excel column to be converted}
}
\description{
Convert charater string of Excel column to a numeric
}
\examples{
excel_column_to_numeric("A")
excel_column_to_numeric("AAA")
}
