% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispensing_days.R, R/get_dispensing_days.R
\name{get_dispensing_days}
\alias{get_dispensing_days}
\title{get_dispensing_days function}
\source{
\url{https://www.gov.uk/bank-holidays.json}

\url{https://www.gov.uk/bank-holidays.json}
}
\usage{
get_dispensing_days(current_year)

get_dispensing_days(current_year)
}
\arguments{
\item{year}{The latest financial year of data required. For example, 2022
for the financial year 2021/22}
}
\description{
This function gets the dispensing days for a time period based on the input year
using bank holiday data from the UK government website's API.
Dispensing days are classed as days that are not a bank holiday or Sunday
in a given month.

This function gets the dispensing days for a time period based on the input year
using bank holiday data from the UK government website's API.
Dispensing days are classed as days that are not a bank holiday or Sunday
in a given month.
}
\examples{
get_dispensing_days(2022)

get_dispensing_days(2022)

}
