% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import latest CSV}
\usage{
import_data(path = "data", ...)
}
\arguments{
\item{path}{file path folder where data is stored. defaults to "data" on
assumption that the function will be used in an Rproj}

\item{...}{additional parameters to be passed to \code{data.table::fread}}
}
\value{
returns a data.frame
}
\description{
Import a CSV file ouputted from \code{mumhquarterly::save_data}. The function uses
the timestamp of the file to determine if it is 'latest'.
}
\examples{
df <- import_data()

df <- import_data("Y:/Official Stats")

tmp <- tempdir()

mumhquarterly::save_data(mtcars, dir = tmp)

df <- import_data(file.path(tmp, "data"), stringsAsFactors = TRUE)
}
