% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoBox_no_border.R
\name{infoBox_no_border}
\alias{infoBox_no_border}
\title{Build infoBox with a border}
\usage{
infoBox_no_border(
  header = "Header here",
  text = "More text here",
  backgroundColour = "#005EB8",
  width = "31\%",
  fontColour = "white"
)
}
\arguments{
\item{header}{The header text (h4) to be given to the infoBox}

\item{text}{The further text needed for the infoBox}

\item{backgroundColour}{The fill colour of the infoBox}

\item{width}{The width of the infoBox (e.g "33\%" for 3 column layout)}

\item{fontColour}{The font colour to use in the infoBox}
}
\description{
Create an info box with solid colour to us in a R markdown file
}
\examples{
`r infoBox_no_border()`
}
