% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ons_national_pop.R
\name{ons_national_pop}
\alias{ons_national_pop}
\title{National population reference file import function}
\source{
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatestimeseriesdataset/current/}
}
\usage{
ons_national_pop(
 
    url = "https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/populationestimatestimeseriesdataset/current/pop.csv",
  year,
  area
)
}
\arguments{
\item{url}{The url for the population data to be imported. Defaults to ONS source}

\item{year}{The year/s to filter population by, should be a vector}

\item{area}{The geographical area/s to filter population by, should be a vector}
}
\description{
This function gets the latest mid-year population estimates from ONS,
and imports it in a form to be used within MUMH Quarterly package.
There are arguments to specify year and geographical area.
}
\details{
If error thrown by arguments, function shows users possible input values for year and area.
}
\examples{
ons_national_pop(year = 2020, area = ENPOP)
ons_national_pop(year = c(2015, 2017, 2019), area = c(GBPOP, ENPOP))
ons_national_pop(year = (2016:2020), area = UKPOP)

}
