% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sheet.R
\name{write_sheet}
\alias{write_sheet}
\title{Easy helper for 'write_sheet'}
\usage{
write_sheet(filepath, sheetname, title, notes, dataset)
}
\arguments{
\item{filepath}{the file path to save the workbook to}

\item{sheetname}{the name of the blank sheet to write to}

\item{title}{the title of the sheet which will go into cell A1}

\item{notes}{a list object of the notes to be included on the sheet}

\item{dataset}{the name of the dataset to be written to the names sheet}
}
\description{
Write data to selected sheet of base 'wb' created using create_wb function
}
\examples{
write_sheet("C:\\\\test.xlsx",
"test1",
"title1",
c("note1", "note2", "note3", "note4"),
mtcars)
}
